+++ CHANGELOG.md
@@ -40,6 +40,7 @@ The following plugins were added:
 - Util: added an optional appearance type parameter to CreateDoor()
 - Util: AddScript(), AddNSSFile() and RemoveNWNXResourceFile() now support valid custom resource directory aliases.
 - Visibility: added two new visibility types to always show an object regardless of range.
+- Weapon: Good Aim Feat now takes value from ruleset.2da
 
 ### Deprecated
 - Tweaks: `NWNX_TWEAKS_HIDE_DMS_ON_CHAR_LIST` has been deprecated, use `NWNX_TWEAKS_HIDE_PLAYERS_ON_CHAR_LIST` now
@@ -58,6 +59,7 @@ N/A
 - Rename: community name only obfuscates once a server reset
 - Rename: properly updates original name if NWNX_Creature_SetOriginalName() is used
 - Weapon: fixed bug in SetGreaterWeaponFocusFeat()
+- Weapon: fixed bug in offhand attack and damage bonus calculations with Greater Weapon feats and two handed weapons
 
 ## 8193.16
 https://github.com/nwnxee/unified/compare/build8193.13...build8193.16